requirements.txt
python-telegram-bot
gTTS~=2.2.1
Pillow~=8.1.0
pandas~=1.2.1
numpy~=1.19.5
psycopg2~=2.8.6
python-dotenv~=0.15.0
python-telegram-bot-calendar
newrelic
requests~=2.25.1
pytz~=2020.5
google-api-python-client
google-auth-oauthlib
docx2pdf
telegram